move ContextProviders to App
